Restore Scintilla header
which make x64 release binary compiled with PlatformToolset v141 crash on startup (x64 debug binary works fine though).

+++ scintilla/include/Sci_Position.h
@@ -20,7 +20,7 @@ typedef size_t Sci_PositionU;
 
 
 // For Sci_CharacterRange  which is defined as long to be compatible with Win32 CHARRANGE
-typedef intptr_t Sci_PositionCR;
+typedef long Sci_PositionCR;
 
 #ifdef _WIN32
 	#define SCI_METHOD __stdcall
